Refactor to separate backbone
