+++ keras_cv/models/__init__.py
@@ -94,7 +94,9 @@ from keras_cv.models.mobilenet_v3 import MobileNetV3Small
 from keras_cv.models.object_detection.faster_rcnn.faster_rcnn import FasterRCNN
 from keras_cv.models.object_detection.retina_net.retina_net import RetinaNet
 from keras_cv.models.object_detection.yolo_v8.yolo_v8 import YOLOV8
-from keras_cv.models.object_detection.yolo_v8.yolo_v8 import YOLOV8Backbone
+from keras_cv.models.object_detection.yolo_v8.yolo_v8_backbone import (
+    YOLOV8Backbone,
+)
 from keras_cv.models.object_detection_3d.center_pillar import (
     MultiHeadCenterPillar,
 )
