+++ keras_cv/models/object_detection/yolo_v8/yolo_v8_presets.py
@@ -13,103 +13,7 @@
 # limitations under the License.
 """YOLOv8 Task presets."""
 
-# TODO(ianstenbit): This should preferably use CSPDarkNet presets eventually
-yolo_v8_backbone_presets_no_weights = {
-    "yolov8_n": {
-        "metadata": {
-            "description": ("YOLOv8_N backbone, pretrained on COCO dataset."),
-        },
-        "class_name": "keras_cv.models>YOLOV8Backbone",
-        "config": {
-            "include_rescaling": True,
-            "input_shape": (None, None, 3),
-            "channels": [32, 64, 128, 256],
-            "depths": [1, 2, 2, 1],
-        },
-    },
-    "yolov8_s": {
-        "metadata": {
-            "description": ("YOLOv8_S backbone, pretrained on COCO dataset."),
-        },
-        "class_name": "keras_cv.models>YOLOV8Backbone",
-        "config": {
-            "include_rescaling": True,
-            "input_shape": (None, None, 3),
-            "channels": [64, 128, 256, 512],
-            "depths": [1, 2, 2, 1],
-        },
-    },
-    "yolov8_m": {
-        "metadata": {
-            "description": ("YOLOv8_M backbone, pretrained on COCO dataset."),
-        },
-        "class_name": "keras_cv.models>YOLOV8Backbone",
-        "config": {
-            "include_rescaling": True,
-            "input_shape": (None, None, 3),
-            "channels": [96, 192, 384, 768],
-            "depths": [2, 4, 4, 2],
-        },
-    },
-    "yolov8_l": {
-        "metadata": {
-            "description": ("YOLOv8_L backbone, pretrained on COCO dataset."),
-        },
-        "class_name": "keras_cv.models>YOLOV8Backbone",
-        "config": {
-            "include_rescaling": True,
-            "input_shape": (None, None, 3),
-            "channels": [128, 256, 512, 512],
-            "depths": [3, 6, 6, 3],
-        },
-    },
-    "yolov8_x": {
-        "metadata": {
-            "description": ("YOLOv8_X backbone, pretrained on COCO dataset."),
-        },
-        "class_name": "keras_cv.models>YOLOV8Backbone",
-        "config": {
-            "include_rescaling": True,
-            "input_shape": (None, None, 3),
-            "channels": [160, 320, 640, 640],
-            "depths": [3, 6, 6, 3],
-        },
-    },
-    "yolov8_x6": {
-        "metadata": {
-            "description": ("YOLOv8_X6 backbone, pretrained on COCO dataset."),
-        },
-        "class_name": "keras_cv.models>YOLOV8Backbone",
-        "config": {
-            "include_rescaling": True,
-            "input_shape": (None, None, 3),
-            "channels": [160, 320, 640, 640, 640],
-            "depths": [3, 6, 6, 3, 3],
-        },
-    },
-}
-
-yolo_v8_backbone_presets_with_weights = {
-    "yolov8_n_coco": {
-        "metadata": {"description": ("TODO: ianstenbit")},
-        "class_name": "keras_cv.models>YOLOV8Backbone",
-        "config": yolo_v8_backbone_presets_no_weights["yolov8_n"]["config"],
-        "weights_url": "https://storage.googleapis.com/keras-cv/models/yolov8/coco/yolov8_n_backbone.h5",
-        "weights_hash": "636ba3cba064c7c457e30a0e4759716006c305c30876df1c1caf2e56b99eab6c",
-    },
-    "yolov8_s_coco": {
-        "metadata": {"description": ("TODO: ianstenbit")},
-        "class_name": "keras_cv.models>YOLOV8Backbone",
-        "config": yolo_v8_backbone_presets_no_weights["yolov8_s"]["config"],
-        "weights_url": "https://storage.googleapis.com/keras-cv/models/yolov8/coco/yolov8_s_backbone.h5",
-        "weights_hash": "49ab5da87d6b36a1943e7f111a1960355171332c25312b6cc01526baaecf1b69",
-    },
-}
-
-yolo_v8_backbone_presets = {
-    **yolo_v8_backbone_presets_no_weights,
-    **yolo_v8_backbone_presets_with_weights,
-}
+from keras_cv.models.object_detection.yolo_v8 import yolo_v8_backbone_presets
 
 yolo_v8_presets = {
     "yolov8_n_coco": {
@@ -117,11 +21,14 @@ yolo_v8_presets = {
             "description": ("TODO(ianstenbit): describe YOLOv8_N"),
         },
         "config": {
-            "backbone": yolo_v8_backbone_presets["yolov8_n"],
+            "backbone": yolo_v8_backbone_presets.backbone_presets[
+                "yolov8_n_backbone"
+            ],
             "num_classes": 80,
             "fpn_depth": 1,
             "temp_use_hacky_encoding": True,
         },
+        "class_name": "keras_cv.models>YOLOV8",
         "weights_url": "https://storage.googleapis.com/keras-cv/models/yolov8/coco/yolov8_n.h5",
         "weights_hash": "2b96bd128a70a67a7226496319c1d4d8e33335e551b71b5a726bdc854b65e888",
     },
@@ -130,7 +37,9 @@ yolo_v8_presets = {
             "description": ("TODO(ianstenbit): describe YOLOv8_S"),
         },
         "config": {
-            "backbone": yolo_v8_backbone_presets["yolov8_s"],
+            "backbone": yolo_v8_backbone_presets.backbone_presets[
+                "yolov8_s_backbone"
+            ],
             "num_classes": 80,
             "fpn_depth": 1,
             "temp_use_hacky_encoding": True,
@@ -143,7 +52,9 @@ yolo_v8_presets = {
             "description": ("TODO(ianstenbit): describe YOLOv8_M"),
         },
         "config": {
-            "backbone": yolo_v8_backbone_presets["yolov8_m"],
+            "backbone": yolo_v8_backbone_presets.backbone_presets[
+                "yolov8_m_backbone"
+            ],
             "num_classes": 80,
             "fpn_depth": 2,
             "temp_use_hacky_encoding": True,
@@ -154,7 +65,9 @@ yolo_v8_presets = {
             "description": ("TODO(ianstenbit): describe YOLOv8_L"),
         },
         "config": {
-            "backbone": yolo_v8_backbone_presets["yolov8_l"],
+            "backbone": yolo_v8_backbone_presets.backbone_presets[
+                "yolov8_l_backbone"
+            ],
             "num_classes": 80,
             "fpn_depth": 3,
             "temp_use_hacky_encoding": True,
@@ -165,7 +78,9 @@ yolo_v8_presets = {
             "description": ("TODO(ianstenbit): describe YOLOv8_x"),
         },
         "config": {
-            "backbone": yolo_v8_backbone_presets["yolov8_x"],
+            "backbone": yolo_v8_backbone_presets.backbone_presets[
+                "yolov8_x_backbone"
+            ],
             "num_classes": 80,
             "fpn_depth": 3,
             "temp_use_hacky_encoding": True,
@@ -176,7 +91,9 @@ yolo_v8_presets = {
             "description": ("TODO(ianstenbit): describe YOLOv8_X6"),
         },
         "config": {
-            "backbone": yolo_v8_backbone_presets["yolov8_x6"],
+            "backbone": yolo_v8_backbone_presets.backbone_presets[
+                "yolov8_x6_backbone"
+            ],
             "num_classes": 80,
             "fpn_depth": 3,
             "temp_use_hacky_encoding": True,
